Refactor: class => file
